# openapi.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Uganda-first professional + dating network. Envelope: `{code:1|0, message, data}`. Auth: Bearer JWT.",
    "title": "LinkUp API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/call/active": {
      "get": {
        "operationId": "get_calls.get_active_calls",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/call/active",
        "tags": [
          "api"
        ]
      }
    },
    "/api/call/history": {
      "get": {
        "operationId": "get_calls.get_call_history",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/call/history",
        "tags": [
          "api"
        ]
      }
    },
    "/api/call/ice-config": {
      "get": {
        "operationId": "get_calls.get_ice_config",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/call/ice-config",
        "tags": [
          "api"
        ]
      }
    },
    "/api/call/log": {
      "post": {
        "operationId": "post_calls.create_call_log",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/call/log",
        "tags": [
          "api"
        ]
      }
    },
    "/api/call/log/{call_id}": {
      "put": {
        "operationId": "put_calls.update_call_log",
        "parameters": [
          {
            "in": "path",
            "name": "call_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "PUT /api/call/log/{call_id}",
        "tags": [
          "api"
        ]
      }
    },
...
    "/api/profile/avatar": {
      "post": {
        "operationId": "post_profile.upload_avatar",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/profile/avatar",
        "tags": [
          "api"
        ]
      }
    },
    "/api/profile/change-password": {
      "post": {
        "operationId": "post_profile.change_password",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/profile/change-password",
        "tags": [
          "api"
        ]
      }
    },
    "/api/profile/delete-account": {
      "post": {
        "operationId": "post_profile.delete_account",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/profile/delete-account",
        "tags": [
          "api"
        ]
      }
    },
    "/api/profile/update": {
      "post": {
        "operationId": "post_profile.update",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/profile/update",
        "tags": [
          "api"
        ]
      }
    },
    "/api/profile/update-email": {
      "post": {
        "operationId": "post_profile.update_email",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/profile/update-email",
        "tags": [
          "api"
        ]
      }
    },
    "/api/profile/update-phone": {
      "post": {
        "operationId": "post_profile.update_phone",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/profile/update-phone",
        "tags": [
          "api"
        ]
      }
    },
...
    "/v1/auth/password": {
      "post": {
        "operationId": "post_v1_identity.change_password",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /v1/auth/password",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/profile/me": {
      "delete": {
        "operationId": "delete_v1_profile.delete_account",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "DELETE /v1/profile/me",
        "tags": [
          "profile"
        ]
      },
      "get": {
        "operationId": "get_v1_profile.get_my_profile",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /v1/profile/me",
        "tags": [
          "profile"
        ]
      },
      "put": {
        "operationId": "put_v1_profile.update_my_profile",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "PUT /v1/profile/me",
        "tags": [
          "profile"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:5001"
    }
  ],
...
}